src/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
requests-oauthlib==1.3.1
duckdb==1.1.3  # Update this (latest stable)
dbt-duckdb==1.8.2 # Update this to match dbt-core 1.8+ standards
python-dotenv==1.0.0